backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
sqlalchemy
aiosqlite
apscheduler
gtts
moviepy
Pillow
PyPDF2
python-docx
requests
httpx
bcrypt
